Output speed travelled between points

+++ src/main/scala/io/github/howyp/events.scala
@@ -3,5 +3,5 @@ package io.github.howyp
 import java.time.LocalDateTime
 
 trait SimulationEvent { val robotId: RobotId }
-case class TrafficReport(robotId: RobotId, timestamp: LocalDateTime, speed: Int, condition: TrafficCondition) extends SimulationEvent
+case class TrafficReport(robotId: RobotId, timestamp: LocalDateTime, speed: Double, condition: TrafficCondition) extends SimulationEvent
 case class RobotMoved(robotId: RobotId, newLocation: Location) extends SimulationEvent
